Feat :level 3

+++ Assets/Script/UIScript/UIComplete.cs
@@ -16,7 +16,7 @@ public class UIComplete : UIBase
     }
 
     void NextScene(){
-        GameManager.instance.ChangeScene(GameManager.instance.currentScene + 1);
+        GameManager.instance.ChangeScene((int)GameManager.instance.currentScene + 1);
     }
 
     void ToMainMenu()
